pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["figeno", "figeno.data","figeno.GUI","figeno.GUI.data"]



[project]
name = 'figeno'
version = "0.0.1"
description = 'Package for generating genomics figures.'
readme = 'README.md'
authors = [
  { name = 'Etienne Sollier'},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = '>=3.7'
dependencies = [
  "numpy>=1.10.0",
  "pandas>=1.3.0",
  "vcfpy>=0.13.5",
  "scipy>=1.7.3",
  "statsmodels>=0.13.5",
  "pyyaml>=6.0",
  "matplotlib>=3.4.3",
  "customtkinter>=5.2.1",
  "Pillow>=9.0.0",
  "pysam>=0.16.0",
  "pyBigWig>=0.3.18",
  "cooler>=0.9.1",
  "importlib-resources>=5.0.0"

]

[project.scripts]
figeno = "figeno.GUI.gui:main"
figenomake = "figeno.figenomake:main"

[project.urls]
"Homepage" = "https://github.com/CompEpigen/figeno"
